api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.15.2
python-dotenv==1.1.0
email-validator==2.3.0
